balanced

+++ 14-binary_tree_balance.c
@@ -6,7 +6,7 @@
 */
 int binary_tree_balance(const binary_tree_t *tree)
 {
-	int left = 0, right = 0;
+	size_t left = 0, right = 0;
 
 	if (tree == NULL)
 		return (0);
